+++ src/components/drawing/menu/ColorsPicker.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 // import Row from './Row';
 // import Menu from './Popover';
-import { DrawingState } from '../../../store/slices/drawingSlice2';
+import { DrawingState } from '../../../store/slices/drawingSlice';
 import { Grid2 as Grid, Popover } from '@mui/material';
 import IconButton from '@mui/material/IconButton';
 import CircleIcon from '@mui/icons-material/Circle';
