+++ src/store/actions/drawingActions.tsx
@@ -1,5 +1,5 @@
 import { Logger } from '../../utils/Logger';
-import { drawingActions, DrawingState } from '../slices/drawingSlice';
+import { drawingActions, DrawingState, FabricAction, } from '../slices/drawingSlice';
 import { AppThunk } from '../store';
 
 const logger = new Logger('DrawingActions');
@@ -9,13 +9,18 @@ const logger = new Logger('DrawingActions');
  * 
  * @returns {AppThunk<Promise<void>>} Promise.
  */
-export const enableDrawing = (): AppThunk<Promise<void>> => async (dispatch, getState, { signalingService }): Promise<void> => {
+export const enableDrawing = (): AppThunk<Promise<void>> => async (
+	dispatch, 
+	_getState, 
+	{ signalingService }
+): Promise<void> => {
 	logger.debug('enableDrawing()');
 
 	try {
 		await signalingService.sendRequest('enableDrawing');
 
 		dispatch(drawingActions.enableDrawing());
+
 	} catch (error) {
 		logger.error('enableDrawing() [error:"%o"]', error);
 	}
@@ -29,7 +34,7 @@ export const enableDrawing = (): AppThunk<Promise<void>> => async (dispatch, get
 export const disableDrawing = (): 
 AppThunk<Promise<void>> => async (
 	dispatch,
-	getState,
+	_getState,
 	{ signalingService }
 ): Promise<void> => {
 	logger.debug('disableDrawing()');
@@ -51,7 +56,7 @@ AppThunk<Promise<void>> => async (
 export const setDrawingBgColor = (bgColor: DrawingState['bgColor']): 
 AppThunk<Promise<void>> => async (
 	dispatch,
-	getState,
+	_getState,
 	{ signalingService }
 ): Promise<void> => {
 	logger.debug('setDrawingBgColor()');
@@ -62,5 +67,51 @@ AppThunk<Promise<void>> => async (
 		dispatch(drawingActions.setDrawingBgColor(bgColor));
 	} catch (error) {
 		logger.error('setDrawingBgColor() [error:"%o"]', error);
+	}
+};
+
+/**
+ * This thunk action updates the canvas .
+ * 
+ * @returns {AppThunk<Promise<void>>} Promise.
+ */
+export const updateCanvasState = (action: FabricAction): 
+AppThunk<Promise<void>> => async (
+	dispatch,
+	_getState,
+	{ signalingService }
+): Promise<void> => {
+	logger.debug('updateCanvasState()');
+
+	try {
+		await signalingService.sendRequest('updateCanvasState', action);
+
+		// dispatch(drawingActions.recordAction(action));
+
+	} catch (error) {
+		logger.error('updateCanvasState() [error:"%o"]', error);
+	}
+};
+
+/**
+ * This thunk action clears the canvas .
+ * 
+ * @returns {AppThunk<Promise<void>>} Promise.
+ */
+export const clearCanvas = (): 
+AppThunk<Promise<void>> => async (
+	dispatch,
+	_getState,
+	{ signalingService }
+): Promise<void> => {
+	logger.debug('clearCanvas()');
+
+	try {
+		await signalingService.sendRequest('clearCanvas');
+
+		dispatch(drawingActions.clear);
+
+	} catch (error) {
+		logger.error('clearCanvas() [error:"%o"]', error);
 	}
 };
