+++ src/store/slices/drawingSlice.tsx
@@ -1,40 +1,63 @@
+/* eslint-disable indent */
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { BasicTransformEvent, FabricObject } from 'fabric';
+import { castDraft } from 'immer';
+
+// adds custom prop to fabric objects
+declare module 'fabric' {
+    // eslint-disable-next-line no-shadow
+    interface FabricObject {
+      id?: number;
+    }
+    interface SerializedObjectProps {
+      id?: number;
+    }
+	interface CanvasEvents {
+		'object:start:modify': Partial<BasicTransformEvent> & {
+			target: FabricObject
+		};
+	}
+  }
+FabricObject.customProperties = [ 'id' ];
 
 export interface DrawingState extends HistoryState {
-	drawingEnabled?: boolean
-	zoom: number
-	tools: [ 'pencilBrush', 'text', 'eraser' ]
-	tool: DrawingState['tools'][number]
-	pencilBrushSize: number
-	pencilBrushSizeRange: { min: number, max: number }
-	textSize: number
-	textSizeRange: { min: number, max: number }
-	eraserSize: number,
-	eraserSizeRange: { min: number, max: number }
-	colorsPickers: [ 'Row', 'Popover' ]
-	colorsPicker: DrawingState['colorsPickers'][number]
-	colors: [ 'black', 'white', 'gray', 'green', 'yellow', 'orange', 'red', 'blue', 'purple' ],
-	color: DrawingState['colors'][number]
-	bgColors: ['gray', 'white', 'black']
-	bgColor: DrawingState['bgColors'][number]
+    drawingEnabled?: boolean
+    zoom: number
+    tools: [ 'edit', 'pencilBrush', 'text', 'eraser' ]
+    tool: DrawingState['tools'][number]
+    pencilBrushSize: number
+    pencilBrushSizeRange: { min: number, max: number }
+    textSize: number
+    textSizeRange: { min: number, max: number }
+    eraserSize: number,
+    eraserSizeRange: { min: number, max: number }
+    colorsPickers: [ 'Row', 'Popover' ]
+    colorsPicker: DrawingState['colorsPickers'][number]
+    colors: [ 'black', 'white', 'gray', 'green', 'yellow', 'orange', 'red', 'blue', 'purple' ]
+    color: DrawingState['colors'][number]
+    bgColors: ['white', 'black', 'gray']
+    bgColor: DrawingState['bgColors'][number]
+	updateAction?: FabricAction
+	initiateAction?: FabricAction[]
+	clearAction: boolean
 }
 
 export interface HistoryState {
-	history: {
-		past: FabricAction[]
-		future: FabricAction[]
-	}
+    history: {
+        past: FabricAction[]
+        future: FabricAction[]
+    }
 }
 
 export interface FabricAction {
-	object: fabric.Object['toObject']
-	status: 'added' | 'modified' | 'removed'
+    object: FabricObject
+	status: 'added' | 'modified' | 'removed' | 'editing' | 'lock'
 }
 
-const initialState: DrawingState_ = {
+const initialState: DrawingState = {
 	drawingEnabled: false,
-	tools: [ 'pencilBrush', 'text', 'eraser' ],
-	tool: 'pencilBrush',
+	tools: [ 'edit', 'pencilBrush', 'text', 'eraser' ],
+	tool: 'edit',
 	pencilBrushSize: 20,
 	pencilBrushSizeRange: { min: 1, max: 100 },
 	textSize: 30,
@@ -46,12 +69,13 @@ const initialState: DrawingState_ = {
 	colorsPicker: 'Popover',
 	colors: [ 'black', 'white', 'gray', 'green', 'yellow', 'orange', 'red', 'blue', 'purple' ],
 	color: 'black',
-	bgColors: [ 'gray', 'white', 'black' ],
-	bgColor: 'gray',
+	bgColors: [ 'white', 'black', 'gray' ],
+	bgColor: 'white',
 	history: {	
 		past: [],
 		future: [],
-	}
+	},
+	clearAction: false,
 };
 
 const drawingSlice = createSlice({
@@ -62,6 +86,7 @@ const drawingSlice = createSlice({
 			state.drawingEnabled = true;
 		}),
 		disableDrawing: ((state) => {
+			state.updateAction = undefined;
 			state.drawingEnabled = false;
 		}),
 		setDrawingTool: ((state, action: PayloadAction<DrawingState['tool']>) => {
@@ -78,9 +103,21 @@ const drawingSlice = createSlice({
 		}),
 
 		recordAction: ((state, action: PayloadAction<FabricAction>) => {
-			state.history.past.push(action.payload);
+			state.history.past.push(castDraft(action.payload));
 			state.history.future = [];
 		}),
+
+		updateCanvas: ((state, action: PayloadAction<FabricAction>) => {
+			state.updateAction = castDraft(action.payload);
+		}),
+
+		InitiateCanvas: ((state, action: PayloadAction<FabricAction[]>) => {
+			state.initiateAction = castDraft(action.payload);
+		}),
+
+		clearCanvas: (state) => {
+			state.clearAction = !state.clearAction;
+		},
 
 		undo: (state) => {
 			const previousAction = state.history.past.pop();
@@ -89,6 +126,7 @@ const drawingSlice = createSlice({
 				state.history.future.push(previousAction);
 			}
 		},
+
 		redo: (state) => {
 			const nextAction = state.history.future.pop();
 
@@ -96,6 +134,7 @@ const drawingSlice = createSlice({
 				state.history.past.push(nextAction);
 			}
 		},
+
 		clear: (state) => {
 			state.history.past = [];
 			state.history.future = [];
@@ -104,8 +143,9 @@ const drawingSlice = createSlice({
 		setDrawingZoom: ((state, action: PayloadAction<number>) => {
 			state.zoom = action.payload;
 		}),
+
 		setDrawingPencilBrushSize: ((state, action: PayloadAction<{ operation: 'inc' | 'dec'}>) => {
-			
+            
 			const operation = action.payload.operation;
 			const { min, max } = state.pencilBrushSizeRange;
 			const curr = state.pencilBrushSize;
@@ -113,9 +153,9 @@ const drawingSlice = createSlice({
 			(operation === 'inc' && curr < max) && state.pencilBrushSize++;
 			(operation === 'dec' && curr > min) && state.pencilBrushSize--;
 		}),
-	
+    
 		setDrawingTextSize: ((state, action: PayloadAction<{ operation: 'inc' | 'dec'}>) => {
-			
+            
 			const operation = action.payload.operation;
 			const { min, max } = state.textSizeRange;
 			const curr = state.textSize;
@@ -123,9 +163,9 @@ const drawingSlice = createSlice({
 			(operation === 'inc' && curr < max) && state.textSize++;
 			(operation === 'dec' && curr > min) && state.textSize--;
 		}),
-	
+    
 		setDrawingEraserSize: ((state, action: PayloadAction<{ operation: 'inc' | 'dec'}>) => {
-			
+            
 			const operation = action.payload.operation;
 			const { min, max } = state.eraserSizeRange;
 			const curr = state.eraserSize;
